purpleair-backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
flask-cors==4.0.1
Flask-Limiter==3.8.0
cachetools==5.5.0
gunicorn==23.0.0
gevent==24.2.1
